requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
psycopg2-binary>=2.9
pydantic>=2.5
python-dotenv>=1.0
python-telegram-bot>=21.0
slowapi>=0.1.9
alembic>=1.13
SQLAlchemy>=2.0
